requirements.txt
streamlit
google-auth-oauthlib
google-api-python-client
openai
pandas
requests
beautifulsoup4